% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gaussian.R
\name{build_model}
\alias{build_model}
\title{Build Gaussian Model}
\usage{
build_model(ngaussian = 1)
}
\arguments{
\item{ngaussian}{Numerical. Number of Gaussian component to consider.}
}
\value{
A \code{formula} object.
}
\description{
Build Gaussian Model
}
\examples{
myfunction <- build_model(2)
}

