% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gaussian.R
\name{cdom_gaussian}
\alias{cdom_gaussian}
\title{Gaussian decomposition on CDOM spectra}
\usage{
cdom_gaussian(x, y, filter = TRUE, merge = TRUE, min_distance = 50,
  min_height = 1)
}
\arguments{
\item{x}{Numerical vector of wavelengths.}

\item{y}{Numerical vector of absorption/absorbance values.}

\item{filter}{Logical. Should the spectra filtered? Default is TRUE.}

\item{merge}{Logical. Determines if initially estimated Gaussian components
should be merged.}

\item{min_distance}{Minimum distance in nm allowed between possible Gaussian
components. Default is 50 nm.}

\item{min_height}{Numerical. Determines the minimum height (per meter) to
model Gaussian components. This is usefull to set a threshold to avoid
modeling noise at higher wavelenghts. Default is 1.}
}
\value{
A data frame with estimated parameters.
}
\description{
Gaussian decomposition on CDOM spectra
}
\examples{
\dontrun{cdom_gaussian(x = wl, y = spc, min_distance = 50)}
}

