% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gaussian.R
\name{coef.cdom_gaussian}
\alias{coef.cdom_gaussian}
\title{Extract Model Coefficients}
\usage{
\method{coef}{cdom_gaussian}(object, ...)
}
\arguments{
\item{object}{An object returned by \code{cdom_gaussian()}.}

\item{...}{other arguments.}
}
\value{
Coefficients extracted from the model object \code{object}.
}
\description{
Extract Model Coefficients
}
\examples{
data(spectra)
myfit <- cdom_gaussian(spectra$wavelength,
                       spectra$spc1,
                       min_distance = 25,
                       min_height = 0.5)

coef(myfit)
}

