% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian.R
\name{coef.gaussian_fit}
\alias{coef.gaussian_fit}
\title{Extract Model Coefficients}
\usage{
\method{coef}{gaussian_fit}(object, ...)
}
\arguments{
\item{object}{An object returned by \code{cdom_gaussian()}.}

\item{...}{other arguments.}
}
\value{
Coefficients extracted from the model object \code{object}.
}
\description{
Extract Model Coefficients
}
\examples{
data(spectra)
myfit <- cdom_gaussian(spectra$wavelength,
                       spectra$spc21,
                       min_distance = 50,
                       min_height = 10)

coef(myfit)
}

