% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fit_exponential.R
\name{fit_exponential}
\alias{fit_exponential}
\title{Fit an exponential model to CDOM data.}
\usage{
fit_exponential(wl, spectra, wl0, startwl, endwl)
}
\arguments{
\item{wl}{The wavelength vector.}

\item{spectra}{The spectra vector.}

\item{wl0}{The reference wavelength (ex.: 350).}

\item{startwl}{The starting wavelength (ex.: 240).}

\item{endwl}{The ending wavelength (ex.: 600).}
}
\value{
A list contianing the \code{nls} object and the \code{R2}. NULL if
the model did not converged.
}
\description{
Fit an exponential model to CDOM data.
}
\details{
\deqn{y = a0 + e^{(-S(x - \lambda_0))} + K}
}
\examples{
# Fit an exponential model using the reference wavelength 350 between 190 and 900 nm.

data(spectra)

fit <- fit_exponential(spectra$wavelength, spectra$absorbance, 350, 190, 900)
summary(fit)

plot(spectra$wavelength, spectra$absorbance)
lines(spectra$wavelength, predict(fit), col = "red")
}

