% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gaussian.R
\name{fitCDOMcomponents}
\alias{fitCDOMcomponents}
\title{Gaussian decomposition on CDOM spectra}
\usage{
fitCDOMcomponents(x, y, filter = TRUE, min_distance)
}
\arguments{
\item{x}{Numerical vector of wavelengths.}

\item{y}{Numerical vector of absorption/absorbance values.}

\item{filter}{Logical. Should the spectra filtered? Default is TRUE.}

\item{min_distance}{Minimum distance in nm allowed between possible Gaussian
components.}
}
\value{
A data frame with estimated parameters.
}
\description{
Gaussian decomposition on CDOM spectra
}
\examples{
\dontrun{fitCDOMcomponents(x = wl, y = spc, min_distance = 50)}
}

