% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gaussian.R
\name{plot.cdom_gaussian}
\alias{plot.cdom_gaussian}
\title{Plot CDOM Gaussian components}
\usage{
\method{plot}{cdom_gaussian}(x, ...)
}
\arguments{
\item{x}{An object returned by the \code{cdom_gaussian()} function.}

\item{...}{other arguments.}
}
\value{
A ggplot2 plot.
}
\description{
Plot CDOM Gaussian components
}
\examples{
data(spectra)
myfit <- cdom_gaussian(spectra$wavelength,
                       spectra$spc1,
                       min_distance = 25,
                       min_height = 0.5)
plot(myfit)
}

