% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponential.R
\name{plot.exponential_fit}
\alias{plot.exponential_fit}
\title{Plot a Fitted CDOM Exponential Curve}
\usage{
\method{plot}{exponential_fit}(x, ...)
}
\arguments{
\item{x}{An object returned by \code{cdom_exponential}.}

\item{...}{other arguments.}
}
\value{
A ggplot2 object.
}
\description{
Plot a Fitted CDOM Exponential Curve
}
\examples{
library(ggplot2)

data(spectra)

fit <- cdom_exponential(spectra$wavelength, spectra$spc1, 350, 190, 900)
p <- plot(fit)
p
p + ggtitle("My super fit")
}
