% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian.R
\name{plot.gaussian_fit}
\alias{plot.gaussian_fit}
\title{Plot CDOM Gaussian components}
\usage{
\method{plot}{gaussian_fit}(x, ...)
}
\arguments{
\item{x}{An object returned by the \code{cdom_gaussian()} function.}

\item{...}{other arguments.}
}
\value{
A ggplot2 plot.
}
\description{
Plot CDOM Gaussian components
}
\examples{
data(spectra)
myfit <- cdom_gaussian(spectra$wavelength,
                       spectra$spc21,
                       min_distance = 50,
                       min_height = 10)
plot(myfit)
}

