% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponential.R
\name{predict.exponential_fit}
\alias{predict.exponential_fit}
\title{Predict method for CDOM exponential fit}
\usage{
\method{predict}{exponential_fit}(object, ...)
}
\arguments{
\item{object}{An object returned by \code{cdom_exponential}.}

\item{...}{other arguments.}
}
\value{
A numerical vector with predicted values.
}
\description{
Predict method for CDOM exponential fit
}
\examples{
data(spectra)

fit <- cdom_exponential(spectra$wavelength, spectra$spc1, 350, 190, 900)
predict(fit)
}

