% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian.R
\name{predict.gaussian_fit}
\alias{predict.gaussian_fit}
\title{Gaussian Model Predictions}
\usage{
\method{predict}{gaussian_fit}(object, ...)
}
\arguments{
\item{object}{a \code{cdom_gaussian} model object for which prediction is
desired.}

\item{...}{other arguments.}
}
\value{
A data frame containing
 \describe{
  \item{exponential_part}{The fitted exponential part of the model}
  \item{component1 ... componentn}{The individual Gaussian components}
  \item{spectra}{The fitted spectra}
}
}
\description{
Gaussian Model Predictions
}
\examples{
data(spectra)
myfit <- cdom_gaussian(spectra$wavelength,
                       spectra$spc1,
                       min_distance = 25,
                       min_height = 0.5)

predict(myfit)
}

