% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spectral_slope.R
\name{spectral_slope}
\alias{spectral_slope}
\title{Calculate the spectral slope of CDOM spectra}
\usage{
spectral_slope(wl, spectra, interval = 21, r2threshold = 0.8)
}
\arguments{
\item{wl}{The wavelength vector.}

\item{spectra}{The spectra vector.}

\item{interval}{The interval used to claculate each slope (default = 21 nm).}

\item{r2threshold}{The r2 threshold that determines if a slope is "valide".
The default value is 0.8 meaning that the determination coefficient of the
regression between log-transformed data and wavelength should be >= 0.8.}
}
\value{
A dataframe containing the centered wavelength, the calculated slope
and the determination coefficient of the linear regression used to claculate the slope.
}
\description{
Calculate the spectral slope of CDOM spectra has proposed by
}
\examples{
data(spectra)

res <- spectral_slope(spectra$wavelength, spectra$absorbance)
plot(res$wl, res$s, type = "l")
}
\references{
\url{http://doi.wiley.com/10.4319/lo.2009.54.2.0590}
}

