% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponential.R
\name{summary.exponential_fit}
\alias{summary.exponential_fit}
\title{Summary of a CDOM exponential fit}
\usage{
\method{summary}{exponential_fit}(object, ...)
}
\arguments{
\item{object}{An object returned by \code{cdom_exponential}.}

\item{...}{other arguments.}
}
\value{
A numerical vector with estimated coefficients.
}
\description{
Summary of a CDOM exponential fit
}
\examples{
data(spectra)

fit <- cdom_exponential(spectra$wavelength, spectra$spc1, 350, 190, 900)
summary(fit)
}
