% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spectra.R
\docType{data}
\name{spectra}
\alias{spectra}
\title{CDOM absorbance data.}
\format{A data frame with 711 rows and 26 variables}
\usage{
data(spectra)
}
\description{
Simple absorbance spectra used to test package's functions.
}
\details{
\itemize{
  \item wavelength.  Wavelengths used for measurements (190-900 nm.)
  \item Absorbance.
}
}
\examples{
library(ggplot2)
library(tidyr)
data("spectra")
spectra <- gather(spectra, sample, absorbance, -wavelength)
ggplot(spectra, aes(x = wavelength, y = absorbance, group = sample)) +
 geom_line(size = 0.1)
}
\keyword{datasets}

